+logs to console (slf4j->log4j

+++ src/main/java/com/github/aasten/transportconcurrent/objects/Route.java
@@ -5,7 +5,7 @@ import java.util.Iterator;
 public interface Route {
     interface RouteElement {
         Station nextStation();
-        double distanceMeters();
+        int distanceMeters();
     }
     Iterator<RouteElement> getFirst();
 }
